==> code/1.7.L2_normalize_scale_datasets.ipynb <==
## 1.7. L2 normalize and scale dataset

This notebook can be used to create Anndata and Seurat normalized and scaled count matrices from the complete time series or the control subset. For some downstream analyses (like calculating the pseudotemporal ordering in R script 3.1) we will need a normalized and scaled version of the data. In case you want to run those codes you will need to run this Notebook first.

### loading python libraries

# standard libraries
import numpy as np
import pandas as pd

# single cell libraries
import anndata as ad
import scanpy as sc
sc.settings.verbosity = 0 

# other 
import warnings
warnings.filterwarnings('ignore') #ignore filter warnings
import sklearn as sk #used for L2 normalization
import sklearn.preprocessing #used for L2 normalization
import os #to get absolute path of file

### load R into Notebook 
%load_ext rpy2.ipython 

### Load data

Here, we load the filtered complete dataset and control-only HSPC subset (as filtered in notebook 1.1 and 1.3 respectively).

### load Anndata object with the control subset 
adata_ctrl_path = "../data/count_matrices/control_filtered.h5ad"
adata_ctrl = sc.read_h5ad(adata_ctrl_path)

### load Anndata object with the HSPC dataset (all four subsets incl.)
adata_path = "../data/count_matrices/all_filtered.h5ad"
adata = sc.read_h5ad(adata_path)

### L2 normalize both dataset
adata_ctrl.X = sk.preprocessing.normalize(adata_ctrl.X, norm='l2')
adata.X = sk.preprocessing.normalize(adata.X, norm='l2')

### Save as Anndata objects

Save the L2 normalized version of the control subset and complete HSPC dataset as Anndata objects.

### save Anndata object
name_file = "control_filtered_L2-normalized.h5ad"
adata_ctrl_path  = "../data/count_matrices/" + name_file
adata_ctrl.write_h5ad(adata_ctrl_path)

### save Anndata object
name_file = "all_filtered_L2-normalized.h5ad"
adata_path  = "../data/count_matrices/" + name_file
adata.write_h5ad(adata_path)

### Save as Seurat objects

Convert the L2 normalized version of the control subset and complete HSPC dataset from Anndata to Seurat objects.

### get absolute path of anndata object
adata_ctrl_path = os.path.abspath(adata_ctrl_path)

%%R -i adata_ctrl_path 

### load packages
library(Seurat) 
library(SeuratDisk)

### load dataset
dataset_dir <- adata_ctrl_path

### convert h5ad to h5seurat 
Convert(dataset_dir, paste0(tools::file_path_sans_ext(dataset_dir), ".h5seurat"), overwrite=TRUE)

### get absolute path of anndata object
adata_path = os.path.abspath(adata_path)

%%R -i adata_path 

### load packages
library(Seurat) 
library(SeuratDisk)

### load dataset
dataset_dir <- adata_path

### convert h5ad to h5seurat 
Convert(dataset_dir, paste0(tools::file_path_sans_ext(dataset_dir), ".h5seurat"), overwrite=TRUE)

### Scale data

### scale data
sc.pp.scale(adata_ctrl)
sc.pp.scale(adata)

### Save as Anndata objects

Save the L2 normalized and scaled version of the control subset and complete HSPC dataset as Anndata objects.

### save Anndata object
name_file = "control_filtered_L2-normalized_scaled.h5ad"
adata_ctrl_path  = "../data/count_matrices/" + name_file
adata_ctrl.write_h5ad(adata_ctrl_path)

### save Anndata object
name_file = "all_filtered_L2-normalized_scaled.h5ad"
adata_path  = "../data/count_matrices/" + name_file
adata.write_h5ad(adata_path)

### Save as Seurat objects

Convert the L2 normalized and scaled version of the control subset and complete HSPC dataset from Anndata to Seurat objects.

### get absolute path of anndata object
adata_ctrl_path = os.path.abspath(adata_ctrl_path)

%%R -i adata_ctrl_path 

### load packages
library(Seurat) 
library(SeuratDisk)

### load dataset
dataset_dir <- adata_ctrl_path

### convert h5ad to h5seurat 
Convert(dataset_dir, paste0(tools::file_path_sans_ext(dataset_dir), ".h5seurat"), overwrite=TRUE)

### get absolute path of anndata object
adata_path = os.path.abspath(adata_path)

%%R -i adata_path 

### load packages
library(Seurat) 
library(SeuratDisk)

### load dataset
dataset_dir <- adata_path

### convert h5ad to h5seurat 
Convert(dataset_dir, paste0(tools::file_path_sans_ext(dataset_dir), ".h5seurat"), overwrite=TRUE)
